Backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
gunicorn==24.1.1
idna==3.11
mysqlclient==2.2.7
numpy==2.4.1
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
pillow==12.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
reportlab==4.4.9
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.11.0
